Move from_circle_and_points back to ClineArc to avoid corner cases
